= enhance crypto and jwt autoconfigure. + add lombok config to ignore generates. + add unit test and code comment. ! fix author.

+++ tranquil-components/tranquil-core/src/main/java/com/ycourlee/tranquil/core/util/FileUtil.java
@@ -1,5 +1,6 @@
 package com.ycourlee.tranquil.core.util;
 
+import com.ycourlee.tranquil.core.CommonConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -7,26 +8,37 @@ import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 
 /**
- * @author yoooonn
+ * @author yooonn
  * @date 2021.04.20
  */
 public class FileUtil {
 
     private static final Logger log = LoggerFactory.getLogger(FileUtil.class);
 
-    public void writeToFile(byte[] buffer, String filepath) {
-        File file = new File(filepath);
-        FileOutputStream fileOutputStream = null;
+    private FileUtil() {}
 
+    public static void writeToFile(String text, String filepath) {
+        writeToFile(text.getBytes(StandardCharsets.UTF_8), filepath);
+    }
+
+    public static void writeToFile(byte[] buffer, String filepath) {
+        File file = new File(filepath);
+        mkdirIfNotExist(file);
+        FileOutputStream fileOutputStream = null;
         BufferedOutputStream bufferedOutputStream = null;
         try {
             fileOutputStream = new FileOutputStream(file);
             bufferedOutputStream = new BufferedOutputStream(fileOutputStream);
-            bufferedOutputStream.write(buffer, 0, buffer.length);
+            int idx = 0;
+            while (idx < buffer.length) {
+                int len = Math.min(buffer.length - idx, CommonConstants.BUFFER);
+                bufferedOutputStream.write(buffer, idx, len);
+                idx += len;
+            }
             bufferedOutputStream.flush();
-
         } catch (IOException e) {
             log.error("写入失败, filepath = {}, e = {}", filepath, e.getMessage());
         } finally {
@@ -36,7 +48,6 @@ public class FileUtil {
                 }
                 if (fileOutputStream != null) {
                     fileOutputStream.close();
-
                 }
             } catch (IOException e) {
                 log.error("IO关闭失败, e = {}", e.getMessage());
@@ -44,22 +55,38 @@ public class FileUtil {
         }
     }
 
-    public void rmDirOrFile(String filepath) {
+    private static void mkdirIfNotExist(File file) {
+        if (!file.exists()) {
+            // noinspection ResultOfMethodCallIgnored
+            file.getParentFile().mkdirs();
+        }
+    }
+
+    public static void rmDirOrFile(String filepath) {
         File file = new File(filepath);
         if (!file.exists()) {
             return;
         }
         if (file.isFile()) {
+            // noinspection ResultOfMethodCallIgnored
             file.delete();
-            log.info("delete a file! filepath = {}", filepath);
+            log.trace("delete a file, filepath = {}", filepath);
         } else if (file.isDirectory()) {
             File[] files = file.listFiles();
-            log.info("delete all files in directory = {}, they are output in log", filepath);
+            log.trace("delete all files in directory = {}, they are output in log", filepath);
+            if (files == null) {
+                return;
+            }
             for (File f : files) {
                 rmDirOrFile(f.getAbsolutePath());
             }
+            // noinspection ResultOfMethodCallIgnored
             file.delete();
-            log.info("delete the directory at last");
+            log.trace("delete the directory at last");
         }
+    }
+
+    public static boolean exist(String filepath) {
+        return new File(filepath).exists();
     }
 }
